Laragigs job-board

+++ app/Http/Controllers/ListingController.php
@@ -49,6 +49,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields = $request->validate([$formFields['user_id'] = auth()->id()]);
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully');
@@ -57,7 +59,6 @@ class ListingController extends Controller
     // Show Edit Form
     public function edit(Listing $listing)
     {
-//        dd($listing);
         return view('listings.edit', [
             'listing' => $listing
         ]);
@@ -65,6 +66,11 @@ class ListingController extends Controller
 
     public function update(Request $request, Listing $listing)
     {
+        // Make sure logged in user is osnwer
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -87,8 +93,19 @@ class ListingController extends Controller
     // Delete Listing
     public function destroy(Listing $listing)
     {
+        // Make sure logged in user is osnwer
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // Manage Listing
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 
 }
